Abstract class TestWithDB for per test/class variants
