+++ cloudant-client/src/test/java/com/cloudant/tests/base/TestWithDbPerTest.java
@@ -1,8 +1,5 @@
 package com.cloudant.tests.base;
 
-import com.cloudant.client.api.CloudantClient;
-import com.cloudant.client.api.Database;
-import com.cloudant.tests.extensions.CloudantClientExtension;
 import com.cloudant.tests.extensions.DatabaseExtension;
 import com.cloudant.tests.extensions.MultiExtension;
 
@@ -11,13 +8,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
 // Base class for tests which require a new DB for each test method
-public class TestWithDbPerTest {
+public class TestWithDbPerTest extends TestWithDb {
 
-    protected static CloudantClientExtension clientResource = new CloudantClientExtension();
     protected static DatabaseExtension.PerTest dbResource = new DatabaseExtension.PerTest(clientResource);
-
-    protected static CloudantClient account;
-    protected static Database db;
 
     @RegisterExtension
     protected static MultiExtension perTestExtensions = new MultiExtension(clientResource, dbResource);
